// src/types/generated/TunnelConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TunnelKind } from "./TunnelKind";

export type TunnelConfig = { id: string, serverId: string, name: string, kind: TunnelKind, bindHost: string, bindPort: number, 
/**
 * Unused for dynamic tunnels.
 */
targetHost: string, targetPort: number, autoStart: boolean, };