// src/types/generated/TunnelInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TunnelKind } from "./TunnelKind";

export type TunnelInput = { serverId: string, name: string, kind: TunnelKind, bindHost: string, bindPort: number, targetHost: string, targetPort: number, autoStart: boolean, };