// src/types/generated/TunnelStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppError } from "./AppError";
import type { TunnelState } from "./TunnelState";

export type TunnelStatus = { id: string, state: TunnelState, activeConnections: number, bytesIn: number, bytesOut: number, error: AppError | null, };